+++ lib/octopress-social/disqus.rb
@@ -3,47 +3,62 @@ module Octopress
     module Disqus
       extend self
 
-      def config(site=nil)
-        @config ||= site['disqus_shortname']
+      attr_accessor :url, :config
+
+      DEFAULTS = {
+        'comments_link_text' => 'Comments',
+        'disabled_comments_text' => 'Comments disabled'
+      }
+
+      def set_config(site)
+        @config ||= DEFAULTS.merge(site['disqus'] || {})
       end
 
-      def url(site, item)
-        Social.full_url(site, item)
+      def set_url(site, item)
+        @url = Social.full_url(site, item)
       end
 
       def identifier(site, item)
-        item['disqus_identifier'] || url(site, item)
+        item['disqus_identifier'] || url
       end
 
       def disqus_comments(site, item)
-        %Q{<div id="disqus_thread"></div>
-          <script type="text/javascript">
-            var disqus_shortname = '#{config}';
-            var disqus_url = '#{url(site, item)}';
-            var disqus_identifier = '#{identifier(site, item)}';
-            var disqus_title = '#{item['title']}';
-            #{embed_script('embed')}
-          </script>
-          <noscript>Please enable JavaScript to view the <a href="https://disqus.com/?ref_noscript" rel="nofollow">comments powered by Disqus.</a></noscript>
-        }
-      end
-
-      def disqus_count_script(site, item)
-        %Q{<script type="text/javascript">
-        var disqus_shortname = '#{config}';
-        #{embed_script('count')}
-        </script>  
-        }
+        if item['comments'] != false
+          %Q{<div id="disqus_thread"></div>
+            <script type="text/javascript">
+              var disqus_shortname = '#{config['shortname']}';
+              var disqus_url = '#{url}';
+              var disqus_identifier = '#{identifier(site, item)}';
+              var disqus_title = '#{item['title']}';
+              #{embed_script('embed')}
+            </script>
+            <noscript>Please enable JavaScript to view the <a href="https://disqus.com/?ref_noscript" rel="nofollow">comments powered by Disqus.</a></noscript>
+          }
+        else
+          ''
+        end
       end
 
       def embed_script(script)
         %Q{(function () {
             var s = document.createElement('script'); s.async = true;
             s.type = 'text/javascript';
-            s.src = '//#{config}.disqus.com/#{script}.js';
+            s.src = '//#{config['shortname']}.disqus.com/#{script}.js';
             (document.getElementsByTagName('HEAD')[0] || document.getElementsByTagName('BODY')[0]).appendChild(s);
           }());
         }
+      end
+
+      def disqus_comments_link(site, item)
+        if item['comments'] != false
+          link = (item['context'] == 'page' ? '' : url)
+          link << '#disqus_thread'
+          %Q{<a class="disqus-comments-link" href="#{link}">Comments</a>}
+        elsif !config['disabled_comments_text'].empty?
+          %Q{<span class="disqus-comments-disabled">#{config['disabled_comments_text']}</span>}
+        else
+          ''
+        end
       end
 
       class Tag < Liquid::Tag
@@ -53,13 +68,12 @@ module Octopress
         end
 
         def render(context)
-          item = context[@input] || context['page']
-          site = context['site']
+          site   = context['site']
+          item   = Octopress::Social.item(context, @input)
 
-          config = Octopress::Social::Disqus.config(site)
-          if config && item['comments'] != false
-            Octopress::Social::Disqus.send(@tag, site, item).gsub(/(\s{2,}|\n)/, ' ').strip
-          end
+          Octopress::Social::Disqus.set_config(site)
+          Octopress::Social::Disqus.set_url(site, item)
+          Octopress::Social::Disqus.send(@tag, site, item).gsub(/(\s{2,}|\n)/, ' ').strip
         end
       end
     end