+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 1.2.0 (2015-03-18)
+- New: Added Disqus comments link tag.
+- New: Facebook comments link tag.
+- New: Tags auto-detect context (sharing a post or the current page).
+- Fixed: Hashtags are added on tweet link.
+
 ### 1.1.0 (2015-03-17)
 - New: Added support Disqus comments.
 - New: Disable Facebook comments on any page or post with `comments: false`.